This includes a bundle of commits to update the code, tests, dependencies to make a huge step towards a 1.0 release. This includes:

- support for newer python-acme releases
- handling nonce errors (with updated acme library), fixes #8
- switch to ACME v2 endpoints
- support for wildcard domains
- a basic logging framework, fixes #3
- test against multiple ACME backends (https://github.com/letsencrypt/boulder and https://github.com/letsencrypt/pebble)
- use docker-compose for simple and reliable ACME backend setup
- Refresh supported Python versions (add new released major versions, drop unsupported old ones)

+++ tests/integration/register.sh
@@ -1,10 +1,14 @@
 #!/bin/bash
+if [[ "$1" = 'generic' ]]; then
+    # we do not have any ACME server as backend
+    exit 0
+fi
+if [[ "$1" = 'pebble' ]]; then
+    export ACME_CAFILE=~/build/letsencrypt/pebble/pebble.minica.pem
+fi
 set -ex
 mkdir -p tests/integration/account tests/integration/log
 rm -f tests/integration/account/*
-bin/acme-register --gen-key --register --email test@example.org configs/integration.ini | tee tests/integration/log/register.log
-if grep 'You need to accept the terms of service' tests/integration/log/register.log > /dev/null; then
-    bin/acme-register --accept-terms-of-service=`grep -E --only-matching http://.* tests/integration/log/register.log` configs/integration.ini
-fi
+bin/acme-register --gen-key --register --email integration-$1@travis$$.org --accept-terms-of-service='something' configs/integration-$1.ini
 echo 'final check'
-bin/acme-register configs/integration.ini
+bin/acme-register configs/integration-$1.ini
